requirements.txt
adal==1.2.7
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
azure-common==1.1.28
azure-core==1.35.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==13.0.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==29.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.0.0
azureml-core==1.60.0.post1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
databricks-sdk==0.58.0
distlib==0.3.9
docker==7.1.0
durationpy==0.10
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
h11==0.16.0
humanfriendly==10.0
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpickle==4.1.1
kiwisolver==1.4.8
knack==0.12.0
kubernetes==33.1.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mlflow==3.1.1
mlflow-skinny==3.1.1
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
ndg-httpsclient==0.5.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
pathspec==0.12.1
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
proto-plus==1.26.1
protobuf==6.31.1
pyaml==25.5.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyreadline3==3.5.4
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.4
requests-auth-aws-sigv4==0.7
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.16.0
SecretStorage==3.3.3
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.1
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
waitress==3.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
xgboost==3.0.2
zipp==3.23.0
